+++ tools/main.R
@@ -8,7 +8,8 @@ lapply(c(
   "yarn_install.R",
   "download_preset_fonts.R",
   "update_gfont_info.R",
-  "expand_variables_article_template.R"
+  "expand_variables_article_template.R",
+  "compile_component_sass.R"
 ), function(file) {
   message("Updating: ", file)
   source(file.path("tools", file), local = TRUE)